スライムも倒せるようにした
