+++ Assets/Scripts/Enemies/Ghost.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: a2ebada532c3e9d4cbaae6e5f270d851
+timeCreated: 1481958313
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
